@@ -3,19 +3,20 @@
 	<modelVersion>4.0.0</modelVersion>
 	<groupId>com.sungardas.enhancedsnapshots</groupId>
 	<artifactId>enhancedsnapshots</artifactId>
-	<version>0.0.1</version>
+	<version>0.0.2</version>
 	<packaging>war</packaging>
 
 	<properties>
 		<org.springframework.version>4.1.1.RELEASE</org.springframework.version>
 		<org.springframework.dao.version>2.0.8</org.springframework.dao.version>
 		<springdata.commons>1.9.0.RELEASE</springdata.commons>
+		<amazonaws.sdk.version>1.10.60</amazonaws.sdk.version>
 	</properties>
 
 	<repositories>
 		<repository>
 			<id>opensourceagility-release</id>
-			<url>http://repo.opensourceagility.com/release</url>
+			<url>http://repo.opensourceagility.com/release/</url>
 		</repository>
 		<repository>
 			<id>spring-maven-milestone</id>
@@ -28,18 +29,27 @@
 		</repository>
 	</repositories>
 
-
-
-
 	<dependencies>
 
-	<dependency>
+		<dependency>
+			<groupId>org.apache.httpcomponents</groupId>
+			<artifactId>httpclient</artifactId>
+			<version>4.4.1</version>
+		</dependency>
+
+		<dependency>
+			<groupId>org.apache.commons</groupId>
+			<artifactId>commons-configuration2</artifactId>
+			<version>2.0</version>
+		</dependency>
+
+		<dependency>
 			<groupId>org.kamranzafar</groupId>
 			<artifactId>jtar</artifactId>
 			<version>2.2</version>
 		</dependency>
 
-	<dependency>
+		<dependency>
 			<groupId>commons-cli</groupId>
 			<artifactId>commons-cli</artifactId>
 			<version>1.3</version>
@@ -51,7 +61,7 @@
 			<version>2.4</version>
 		</dependency>
 		
-	<dependency>
+		<dependency>
 			<groupId>org.json</groupId>
 			<artifactId>json</artifactId>
 			<version>20140107</version>
@@ -65,7 +75,7 @@
 		<dependency>
 			<groupId>com.amazonaws</groupId>
 			<artifactId>aws-java-sdk</artifactId>
-			<version>1.10.11</version>
+			<version>${amazonaws.sdk.version}</version>
 		</dependency>
 		<dependency>
 			<groupId>org.apache.tomcat</groupId>
@@ -242,7 +252,7 @@
 		<dependency>
 			<groupId>org.socialsignin</groupId>
 			<artifactId>spring-data-dynamodb</artifactId>
-			<version>1.0.2.RELEASE</version>
+			<version>4.2.1</version>
 		</dependency>
 
 		<dependency>
@@ -381,6 +391,12 @@
 			<artifactId>jasypt</artifactId>
 			<version>1.9.2</version>
 		</dependency>
+		<dependency>
+			<groupId>junit</groupId>
+			<artifactId>junit</artifactId>
+			<version>4.11</version>
+			<scope>test</scope>
+		</dependency>
 
 
 	</dependencies>
@@ -391,8 +407,8 @@
 				<artifactId>maven-compiler-plugin</artifactId>
 				<version>3.1</version>
 				<configuration>
-					<source>1.7</source>
-					<target>1.7</target>
+					<source>1.8</source>
+					<target>1.8</target>
 				</configuration>
 			</plugin>
 			<plugin>
