@@ -1,28 +1,5 @@
 package com.sungardas.enhancedsnapshots.components;
 
-import com.amazonaws.AmazonClientException;
-import com.sungardas.enhancedsnapshots.aws.dynamodb.model.TaskEntry;
-import com.sungardas.enhancedsnapshots.aws.dynamodb.model.WorkerConfiguration;
-import com.sungardas.enhancedsnapshots.aws.dynamodb.repository.TaskRepository;
-import com.sungardas.enhancedsnapshots.exception.EnhancedSnapshotsInterruptedException;
-import com.sungardas.enhancedsnapshots.service.ConfigurationService;
-import com.sungardas.enhancedsnapshots.service.NotificationService;
-import com.sungardas.enhancedsnapshots.service.SDFSStateService;
-import com.sungardas.enhancedsnapshots.service.TaskService;
-import com.sungardas.enhancedsnapshots.tasks.BackupTask;
-import com.sungardas.enhancedsnapshots.tasks.DeleteTask;
-import com.sungardas.enhancedsnapshots.tasks.RestoreTask;
-import com.sungardas.enhancedsnapshots.tasks.Task;
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
-import org.springframework.beans.factory.ObjectFactory;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.DependsOn;
-import org.springframework.stereotype.Service;
-
-import javax.annotation.PostConstruct;
-import javax.annotation.PreDestroy;
 import java.util.Comparator;
 import java.util.List;
 import java.util.Set;
@@ -31,48 +8,68 @@ import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 import java.util.concurrent.TimeUnit;
 
+import javax.annotation.PostConstruct;
+import javax.annotation.PreDestroy;
+
+import com.amazonaws.AmazonClientException;
+import com.sungardas.enhancedsnapshots.aws.dynamodb.model.TaskEntry;
+import com.sungardas.enhancedsnapshots.aws.dynamodb.repository.TaskRepository;
+import com.sungardas.enhancedsnapshots.exception.EnhancedSnapshotsInterruptedException;
+import com.sungardas.enhancedsnapshots.service.NotificationService;
+import com.sungardas.enhancedsnapshots.service.SDFSStateService;
+import com.sungardas.enhancedsnapshots.service.SystemService;
+import com.sungardas.enhancedsnapshots.service.TaskService;
+import com.sungardas.enhancedsnapshots.tasks.executors.TaskExecutor;
+
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.context.annotation.DependsOn;
+import org.springframework.stereotype.Service;
+
 import static com.sungardas.enhancedsnapshots.aws.dynamodb.model.TaskEntry.TaskEntryStatus.ERROR;
 import static com.sungardas.enhancedsnapshots.aws.dynamodb.model.TaskEntry.TaskEntryStatus.RUNNING;
 
+
 @Service
-@DependsOn("CreateAppConfiguration")
+@DependsOn("SystemService")
 public class WorkersDispatcher {
-    private static final Comparator<TaskEntry> taskComparatorByTimeAndPriority = new Comparator<TaskEntry>() {
-        @Override
-        public int compare(TaskEntry o1, TaskEntry o2) {
-            int priority = o2.getPriority() - o1.getPriority();
-            if (priority != 0) {
-                return priority;
-            }
-            return o1.getSchedulerTime().compareTo(o2.getSchedulerTime());
+
+    private static final Comparator<TaskEntry> taskComparatorByTimeAndPriority = (o1, o2) -> {
+        int priority = o2.getPriority() - o1.getPriority();
+        if (priority != 0) {
+            return priority;
         }
+        return o1.getSchedulerTime().compareTo(o2.getSchedulerTime());
     };
+
     @Autowired
-    private ConfigurationService configurationService;
+    private ConfigurationMediator configurationMediator;
     @Autowired
-    private ObjectFactory<BackupTask> backupTaskObjectFactory;
+    @Qualifier("awsBackupVolumeTaskExecutor")
+    private TaskExecutor awsBackupVolumeTaskExecutor;
     @Autowired
-    private ObjectFactory<DeleteTask> deleteTaskObjectFactory;
+    @Qualifier("awsDeleteTaskExecutor")
+    private TaskExecutor awsDeleteTaskExecutor;
     @Autowired
-    private ObjectFactory<RestoreTask> restoreTaskObjectFactory;
+    @Qualifier("awsRestoreVolumeTaskExecutor")
+    private TaskExecutor awsRestoreVolumeTaskExecutor;
     @Autowired
     private TaskService taskService;
     @Autowired
     private SDFSStateService sdfsStateService;
     @Autowired
+    private SystemService systemService;
+    @Autowired
     private TaskRepository taskRepository;
-
     @Autowired
     private NotificationService notificationService;
 
-    @Value("${enhancedsnapshots.polling.rate}")
-    private int pollingRate;
-    private WorkerConfiguration configuration;
     private ExecutorService executor;
 
     @PostConstruct
     private void init() {
-        configuration = configurationService.getWorkerConfiguration();
         executor = Executors.newSingleThreadExecutor();
         executor.execute(new TaskWorker());
     }
@@ -95,8 +92,6 @@ public class WorkersDispatcher {
 
         @Override
         public void run() {
-            String instanceId = configuration.getConfigurationId();
-
             LOGtw.info("Starting worker dispatcher");
             while (true) {
                 if (Thread.interrupted()) {
@@ -104,35 +99,38 @@ public class WorkersDispatcher {
                 }
                 TaskEntry entry = null;
                 try {
-                    Set<TaskEntry> taskEntrySet = sortByTimeAndPriority(taskRepository.findByStatusAndInstanceIdAndRegular(TaskEntry.TaskEntryStatus.QUEUED.getStatus(), instanceId, Boolean.FALSE.toString()));
+                    Set<TaskEntry> taskEntrySet = sortByTimeAndPriority(taskRepository.findByStatusAndRegular(TaskEntry.TaskEntryStatus.QUEUED.getStatus(), Boolean.FALSE.toString()));
                     while (!taskEntrySet.isEmpty()) {
                         entry = taskEntrySet.iterator().next();
 
-                        Task task = null;
                         if (!taskService.isCanceled(entry.getId())) {
                             switch (TaskEntry.TaskEntryType.getType(entry.getType())) {
                                 case BACKUP:
+                                    //TODO remove when sdfscli will allow to expand cache size at runtime
+                                    if (!sdfsStateService.sdfsIsAvailable()) {
+                                        break;
+                                    }
                                     LOGtw.info("Task was identified as backup");
-                                    task = backupTaskObjectFactory.getObject();
-                                    task.setTaskEntry(entry);
+                                    awsBackupVolumeTaskExecutor.execute(entry);
                                     break;
                                 case DELETE: {
                                     LOGtw.info("Task was identified as delete backup");
-                                    task = deleteTaskObjectFactory.getObject();
-                                    task.setTaskEntry(entry);
+                                    awsDeleteTaskExecutor.execute(entry);
                                     break;
                                 }
                                 case RESTORE:
+                                    if (!sdfsStateService.sdfsIsAvailable()) {
+                                        break;
+                                    }
                                     LOGtw.info("Task was identified as restore");
-                                    task = restoreTaskObjectFactory.getObject();
-                                    task.setTaskEntry(entry);
+                                    awsRestoreVolumeTaskExecutor.execute(entry);
                                     break;
                                 case SYSTEM_BACKUP: {
                                     LOGtw.info("Task was identified as system backup");
                                     notificationService.notifyAboutTaskProgress(entry.getId(), "System backup started", 0);
                                     entry.setStatus(RUNNING.getStatus());
                                     taskRepository.save(entry);
-                                    sdfsStateService.backupState(entry.getId());
+                                    systemService.backup(entry.getId());
                                     taskRepository.delete(entry);
                                     notificationService.notifyAboutTaskProgress(entry.getId(), "System backup finished", 100);
                                     break;
@@ -144,15 +142,16 @@ public class WorkersDispatcher {
                         } else {
                             LOGtw.debug("Task canceled: {}", entry);
                         }
-                        if (task != null) {
-                            task.execute();
-                        }
-                        taskEntrySet = sortByTimeAndPriority(taskRepository.findByStatusAndInstanceIdAndRegular(TaskEntry.TaskEntryStatus.QUEUED.getStatus(), instanceId, Boolean.FALSE.toString()));
+                        taskEntrySet = sortByTimeAndPriority(taskRepository.findByStatusAndRegular(TaskEntry.TaskEntryStatus.QUEUED.getStatus(), Boolean.FALSE.toString()));
                     }
                     sleep();
                 } catch (AmazonClientException e) {
                     LOGtw.error(e);
                 } catch (EnhancedSnapshotsInterruptedException e) {
+                    return;
+                    // this is required to close thread when uninstalling system
+                } catch (IllegalStateException e) {
+                    LOGtw.warn("Stopping worker dispatcher ...");
                     return;
                 } catch (Exception e) {
                     LOGtw.error(e);
@@ -166,7 +165,7 @@ public class WorkersDispatcher {
 
         private void sleep() {
             try {
-                TimeUnit.MILLISECONDS.sleep(pollingRate);
+                TimeUnit.MILLISECONDS.sleep(configurationMediator.getWorkerDispatcherPollingRate());
             } catch (InterruptedException e) {
                 e.printStackTrace();
             }
