@@ -5,29 +5,26 @@ import com.amazonaws.regions.Region;
 import com.amazonaws.regions.Regions;
 import com.amazonaws.services.dynamodbv2.AmazonDynamoDB;
 import com.amazonaws.services.dynamodbv2.AmazonDynamoDBClient;
+import com.amazonaws.services.dynamodbv2.datamodeling.DynamoDBMapper;
+import com.amazonaws.services.dynamodbv2.datamodeling.DynamoDBMapperConfig;
 import com.amazonaws.services.ec2.AmazonEC2;
 import com.amazonaws.services.ec2.AmazonEC2Client;
 import com.amazonaws.services.s3.AmazonS3;
 import com.amazonaws.services.s3.AmazonS3Client;
+import com.amazonaws.util.EC2MetadataUtils;
 import com.sungardas.enhancedsnapshots.components.RetryInterceptor;
+
 import org.socialsignin.spring.data.dynamodb.repository.config.EnableDynamoDBRepositories;
 import org.springframework.aop.framework.ProxyFactoryBean;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Profile;
 
 @Configuration
 @Profile("prod")
-@EnableDynamoDBRepositories(basePackages = "com.sungardas.enhancedsnapshots.aws.dynamodb.repository")
+@EnableDynamoDBRepositories(basePackages = "com.sungardas.enhancedsnapshots.aws.dynamodb.repository", dynamoDBMapperConfigRef = "dynamoDBMapperConfig")
 public class AmazonConfigProvider {
     private InstanceProfileCredentialsProvider  credentialsProvider;
-
-    @Value("${sungardas.worker.configuration}")
-    private String instanceId;
-
-    @Value("${amazon.aws.region}")
-    private String region;
 
     @Bean(name = "retryInterceptor")
     public RetryInterceptor retryInterceptor() {
@@ -48,7 +45,6 @@ public class AmazonConfigProvider {
 
         proxyFactoryBean.setTarget(amazonDynamoDB());
         proxyFactoryBean.setInterceptorNames("retryInterceptor");
-
         return proxyFactoryBean;
     }
 
@@ -72,24 +68,55 @@ public class AmazonConfigProvider {
         return proxyFactoryBean;
     }
 
-    private AmazonDynamoDB amazonDynamoDB() {
+    @Bean
+    public DynamoDBMapperConfig dynamoDBMapperConfig() {
+        DynamoDBMapperConfig.Builder builder = new DynamoDBMapperConfig.Builder();
+        builder.withTableNameOverride(DynamoDBMapperConfig.TableNameOverride.
+                withTableNamePrefix(getDynamoDbPrefix()));
+        return builder.build();
+    }
+
+    @Bean
+    public ProxyFactoryBean amazonDynamoDbMapperProxy() {
+        ProxyFactoryBean proxyFactoryBean = new ProxyFactoryBean();
+
+        proxyFactoryBean.setTarget(dynamoDBMapper());
+        proxyFactoryBean.setInterceptorNames("retryInterceptor");
+
+        return proxyFactoryBean;
+    }
+
+    @Bean(name = "dynamoDB")
+    public AmazonDynamoDB amazonDynamoDB() {
         AmazonDynamoDB amazonDynamoDB = new AmazonDynamoDBClient(amazonCredentialsProvider());
-        amazonDynamoDB.setRegion(Region.getRegion(Regions.fromName(region)));
+        amazonDynamoDB.setRegion(Regions.getCurrentRegion());
         return amazonDynamoDB;
+    }
+
+    private DynamoDBMapper dynamoDBMapper() {
+        return new DynamoDBMapper(amazonDynamoDB(), dynamoDBMapperConfig());
     }
 
     private AmazonEC2 amazonEC2() {
         AmazonEC2 amazonEC2 = new AmazonEC2Client(amazonCredentialsProvider());
-        amazonEC2.setRegion(Region.getRegion(Regions.fromName(region)));
+        amazonEC2.setRegion(Regions.getCurrentRegion());
         return amazonEC2;
     }
 
     private AmazonS3 amazonS3() {
         AmazonS3 amazonS3 = new AmazonS3Client(amazonCredentialsProvider());
-        Region current = Region.getRegion(Regions.fromName(region));
+        Region current = Regions.getCurrentRegion();
         if (!current.equals(Region.getRegion(Regions.US_EAST_1))) {
             amazonS3.setRegion(current);
         }
         return amazonS3;
+    }
+
+    public static String getDynamoDbPrefix() {
+        return getDynamoDbPrefix(EC2MetadataUtils.getInstanceId());
+    }
+
+    public static String getDynamoDbPrefix(String instanceId) {
+        return "ENHANCEDSNAPSHOTS_" + instanceId + "_";
     }
 }
