@@ -5,7 +5,6 @@ import com.amazonaws.services.dynamodbv2.datamodeling.DynamoDBHashKey;
 import com.amazonaws.services.dynamodbv2.datamodeling.DynamoDBIgnore;
 import com.amazonaws.services.dynamodbv2.datamodeling.DynamoDBTable;
 import com.amazonaws.util.json.Jackson;
-import org.json.JSONObject;
 
 
 @DynamoDBTable(tableName = "Tasks")
@@ -38,9 +37,6 @@ public class TaskEntry {
     @DynamoDBAttribute(attributeName = "schedulerTime")
     private String schedulerTime;
 
-    @DynamoDBAttribute(attributeName = "instanceId")
-    private String instanceId;
-
     @DynamoDBAttribute(attributeName = "options")
     private String options;
 
@@ -56,52 +52,17 @@ public class TaskEntry {
     @DynamoDBAttribute
     private String expirationDate;
 
-    public TaskEntry() {
-        super();
-    }
+    @DynamoDBAttribute
+    private String tempVolumeType;
 
-    @Deprecated
-    public TaskEntry(String priority, String status, String type, String volume,
-                     String schedulerManual, String schedulerName, String schedulerTime, String instanceId) {
-        this(priority, instanceId, status, type, volume, schedulerManual, schedulerName, schedulerTime, instanceId, null);
-    }
+    @DynamoDBAttribute
+    private String restoreVolumeType;
 
-    @Deprecated
-    public TaskEntry(String priority, String worker, String status, String type, String volume,
-                     String schedulerManual, String schedulerName, String schedulerTime, String instanceId,
-                     String options) {
-        this.priority = Integer.parseInt(priority);
-        this.worker = worker;
-        this.status = status;
-        this.type = type;
-        this.volume = volume;
-        this.schedulerManual = schedulerManual;
-        this.schedulerName = schedulerName;
-        this.schedulerTime = schedulerTime;
-        this.instanceId = instanceId;
-        this.options = options;
-    }
+    @DynamoDBAttribute
+    private int tempVolumeIopsPerGb;
 
-    public TaskEntry(JSONObject jsonTask) {
-        this();
-
-        try {
-            this.setPriority(jsonTask.getInt("priority"));
-        } catch (RuntimeException emptyPriority) {
-            this.setPriority(0);
-        }
-        setStatus(jsonTask.getString("status"));
-        setType(jsonTask.getString("type"));
-        setVolume(jsonTask.optString("volume"));
-        setSchedulerManual(jsonTask.getBoolean("schedulerManual"));
-        setSchedulerName(jsonTask.optString("schedulerName"));
-        setSchedulerTime(jsonTask.optString("schedulerTime"));
-        setId(jsonTask.getString("id"));
-        setWorker(jsonTask.getString("worker"));
-        setInstanceId(jsonTask.getString("instanceId"));
-        setOptions(jsonTask.optString("options"));
-    }
-
+    @DynamoDBAttribute
+    private int restoreVolumeIopsPerGb;
 
     public String getId() {
         return id;
@@ -114,6 +75,7 @@ public class TaskEntry {
     public int getPriority() {
         return priority;
     }
+
 
     public void setPriority(int priority) {
         this.priority = priority;
@@ -179,14 +141,6 @@ public class TaskEntry {
         this.schedulerTime = schedulerTime;
     }
 
-    public String getInstanceId() {
-        return instanceId;
-    }
-
-    public void setInstanceId(String instanceId) {
-        this.instanceId = instanceId;
-    }
-
     public String getOptions() {
         return options;
     }
@@ -240,16 +194,48 @@ public class TaskEntry {
         return options.split(", ")[0];
     }
 
+    public int getTempVolumeIopsPerGb() {
+        return tempVolumeIopsPerGb;
+    }
+
+    public void setTempVolumeIopsPerGb(int tempVolumeIopsPerGb) {
+        this.tempVolumeIopsPerGb = tempVolumeIopsPerGb;
+    }
+
+    public String getTempVolumeType() {
+        return tempVolumeType;
+    }
+
+    public void setTempVolumeType(String tempVolumeType) {
+        this.tempVolumeType = tempVolumeType;
+    }
+
     @DynamoDBIgnore
     public String getAvailabilityZone() {
         return options.split(", ")[1];
     }
 
-
     @Deprecated
     @Override
     public String toString() {
         return Jackson.toJsonString(this);
+    }
+
+    public String getRestoreVolumeType() {
+        return restoreVolumeType;
+    }
+
+
+    public void setRestoreVolumeType(String restoreVolumeType) {
+        this.restoreVolumeType = restoreVolumeType;
+    }
+
+    public int getRestoreVolumeIopsPerGb() {
+        return restoreVolumeIopsPerGb;
+    }
+
+    public void setRestoreVolumeIopsPerGb(int restoreVolumeIopsPerGb) {
+        this.restoreVolumeIopsPerGb = restoreVolumeIopsPerGb;
     }
 
     public enum TaskEntryType {
