@@ -1,9 +1,10 @@
 package com.sungardas.enhancedsnapshots.aws.dynamodb.repository;
 
-import com.sungardas.enhancedsnapshots.aws.dynamodb.model.WorkerConfiguration;
+import com.sungardas.enhancedsnapshots.aws.dynamodb.model.Configuration;
 import org.socialsignin.spring.data.dynamodb.repository.EnableScan;
 import org.springframework.data.repository.CrudRepository;
 
 @EnableScan
-public interface WorkerConfigurationRepository extends CrudRepository <WorkerConfiguration, String>{
+public interface ConfigurationRepository extends CrudRepository <Configuration, String>{
 }
+
