@@ -1,14 +1,15 @@
 package com.sungardas.enhancedsnapshots.dto;
 
-import com.fasterxml.jackson.annotation.JsonProperty;
-
 import java.util.Collections;
 import java.util.List;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
 
 public class InitConfigurationDto {
     private List<S3> s3 = Collections.EMPTY_LIST;
     private SDFS sdfs;
     private DB db;
+    private String immutableBucketNamePrefix;
 
     public List<S3> getS3() {
         return s3;
@@ -65,12 +66,50 @@ public class InitConfigurationDto {
         }
     }
 
+    public String getImmutableBucketNamePrefix() {
+        return immutableBucketNamePrefix;
+    }
+
+    public void setImmutableBucketNamePrefix(String immutableBucketNamePrefix) {
+        this.immutableBucketNamePrefix = immutableBucketNamePrefix;
+    }
+
     public static class SDFS {
         @JsonProperty("isCreated")
         private boolean created;
         private String volumeName;
         private String volumeSize;
         private String mountPoint;
+        private String minVolumeSize;
+        private String maxVolumeSize;
+        private int sdfsLocalCacheSize;
+        private int maxSdfsLocalCacheSize;
+        private int minSdfsLocalCacheSize;
+
+        public int getSdfsLocalCacheSize() {
+            return sdfsLocalCacheSize;
+        }
+
+        public void setSdfsLocalCacheSize(int sdfsLocalCacheSize) {
+            this.sdfsLocalCacheSize = sdfsLocalCacheSize;
+        }
+
+        public int getMaxSdfsLocalCacheSize() {
+            return maxSdfsLocalCacheSize;
+        }
+
+        public void setMaxSdfsLocalCacheSize(int maxSdfsLocalCacheSize) {
+            this.maxSdfsLocalCacheSize = maxSdfsLocalCacheSize;
+        }
+
+        public int getMinSdfsLocalCacheSize() {
+            return minSdfsLocalCacheSize;
+        }
+
+        public void setMinSdfsLocalCacheSize(int minSdfsLocalCacheSize) {
+            this.minSdfsLocalCacheSize = minSdfsLocalCacheSize;
+        }
+
 
         public boolean isCreated() {
             return created;
@@ -102,6 +141,22 @@ public class InitConfigurationDto {
 
         public void setMountPoint(String mountPoint) {
             this.mountPoint = mountPoint;
+        }
+
+        public String getMinVolumeSize() {
+            return minVolumeSize;
+        }
+
+        public void setMinVolumeSize(final String minVolumeSize) {
+            this.minVolumeSize = minVolumeSize;
+        }
+
+        public String getMaxVolumeSize() {
+            return maxVolumeSize;
+        }
+
+        public void setMaxVolumeSize(final String maxVolumeSize) {
+            this.maxVolumeSize = maxVolumeSize;
         }
     }
 
