@@ -1,11 +1,12 @@
 package com.sungardas.enhancedsnapshots.service;
 
+import java.util.List;
+
 import com.amazonaws.services.ec2.model.AvailabilityZone;
 import com.amazonaws.services.ec2.model.Instance;
 import com.amazonaws.services.ec2.model.Snapshot;
 import com.amazonaws.services.ec2.model.Volume;
-
-import java.util.List;
+import com.amazonaws.services.ec2.model.VolumeType;
 
 public interface AWSCommunicationService {
 
@@ -17,15 +18,16 @@ public interface AWSCommunicationService {
 
     Snapshot waitForCompleteState(Snapshot snapshot);
 
-    Snapshot syncSnapshot(Snapshot snapshot);
+    Snapshot syncSnapshot(String snapshotId);
 
     Volume waitForAvailableState(Volume volume);
 
     Volume syncVolume(Volume volume);
 
-    Volume createVolumeFromSnapshot(String snapshotId, String availabilityZoneName);
-
-    Volume createVolumeFromSnapshot(Snapshot sourceSnapshot, String availabilityZoneName);
+    /**
+     * iopsPerGb paramenter is only required for io1 volume type, for other volume types it will be skipped
+     */
+    Volume createVolumeFromSnapshot(String snapshotId, String availabilityZoneName, VolumeType type, int iopsPerGb);
 
     void deleteVolume(Volume volume);
 
@@ -41,13 +43,9 @@ public interface AWSCommunicationService {
 
     void deleteTemporaryTag(String resourceId);
 
-    Volume createVolume(int size, int iiops, String type);
+    Volume createVolume(int size, VolumeType type);
 
-    Volume createIO1Volume(int size, int iops);
-
-    Volume createGP2Volume(int size);
-
-    Volume createStandardVolume(int size);
+    Volume createIO1Volume(int size, int iopsPerGb);
 
     Instance getInstance(String instanceId);
 
@@ -56,4 +54,17 @@ public interface AWSCommunicationService {
     void setResourceName(String resourceid, String value);
 
     void addTag(String resourceId, String name, String value);
+
+    boolean snapshotExists(String snapshotId);
+
+    boolean volumeExists(String volumeId);
+
+    Snapshot getSnapshot(String snapshotId);
+
+    void dropS3Bucket(String bucketName);
+
+    //TODO: awslog restarted after each sdfs remount
+    //TODO: need to clarify whether we still need restart it
+    //TODO: in case we need, it should be triggered by SystemService while restarting sdfs, not by SDFSStateService
+    void restartAWSLogService();
 }
