@@ -14,10 +14,10 @@ import com.sungardas.enhancedsnapshots.aws.dynamodb.model.BackupEntry;
 import com.sungardas.enhancedsnapshots.aws.dynamodb.model.RetentionEntry;
 import com.sungardas.enhancedsnapshots.aws.dynamodb.repository.BackupRepository;
 import com.sungardas.enhancedsnapshots.aws.dynamodb.repository.RetentionRepository;
+import com.sungardas.enhancedsnapshots.components.ConfigurationMediator;
 import com.sungardas.enhancedsnapshots.dto.RetentionDto;
 import com.sungardas.enhancedsnapshots.exception.DataAccessException;
 import com.sungardas.enhancedsnapshots.service.BackupService;
-import com.sungardas.enhancedsnapshots.service.ConfigurationService;
 import com.sungardas.enhancedsnapshots.service.RetentionService;
 import com.sungardas.enhancedsnapshots.service.SchedulerService;
 import com.sungardas.enhancedsnapshots.service.Task;
@@ -27,7 +27,6 @@ import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.joda.time.DateTime;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 
 import static com.sungardas.enhancedsnapshots.dto.converter.RetentionConverter.toDto;
@@ -58,17 +57,14 @@ public class RetentionServiceImpl implements RetentionService {
     private SchedulerService schedulerService;
 
     @Autowired
-    private ConfigurationService configurationService;
+    private ConfigurationMediator configurationMediator;
 
     private String instanceId;
 
-    @Value("${enhancedsnapshots.retention.cron}")
-    private String cronExpression;
-
     @PostConstruct
     private void init() {
-        instanceId = configurationService.getWorkerConfiguration().getConfigurationId();
-        schedulerService.addTask(getJob(this), cronExpression);
+        instanceId = configurationMediator.getConfigurationId();
+        schedulerService.addTask(getJob(this), configurationMediator.getRetentionCronExpression());
         try {
             apply();
         } catch (AmazonClientException e) {
@@ -79,11 +75,7 @@ public class RetentionServiceImpl implements RetentionService {
     @Override
     public void putRetention(RetentionDto retentionDto) {
         if (volumeService.volumeExists(retentionDto.getVolumeId())) {
-            RetentionEntry entry = toEntry(retentionDto);
-
-            entry.setInstanceId(instanceId);
-
-            retentionRepository.save(entry);
+            retentionRepository.save(toEntry(retentionDto));
             apply();
         } else {
             LOG.error("Volume with id: {} not found", retentionDto.getVolumeId());
@@ -94,7 +86,7 @@ public class RetentionServiceImpl implements RetentionService {
     @Override
     public RetentionDto getRetentionDto(String volumeId) {
         try {
-            return toDto(retentionRepository.findOne(RetentionEntry.getId(volumeId, instanceId)));
+            return toDto(retentionRepository.findOne(volumeId));
         } catch (Exception e) {
             if (volumeService.volumeExists(volumeId)) {
                 return new RetentionDto(volumeId);
@@ -116,7 +108,7 @@ public class RetentionServiceImpl implements RetentionService {
             RetentionEntry retentionEntry = retentions.get(entry.getKey());
             if (retentionEntry != null) {
                 if (isEmpty(retentionEntry)) {
-                    retentionRepository.delete(RetentionEntry.getId(retentionEntry.getVolumeId(), retentionEntry.getInstanceId()));
+                    retentionRepository.delete(retentionEntry.getVolumeId());
                 } else {
                     BackupEntry[] values = entry.getValue().toArray(new BackupEntry[entry.getValue().size()]);
                     applySizeRetention(backupsToRemove, values, retentionEntry);
@@ -128,7 +120,7 @@ public class RetentionServiceImpl implements RetentionService {
 
         for (Map.Entry<String, RetentionEntry> entry : retentions.entrySet()) {
             if (!backups.containsKey(entry.getKey())) {
-                retentionRepository.delete(RetentionEntry.getId(entry.getValue().getVolumeId(), entry.getValue().getInstanceId()));
+                retentionRepository.delete(entry.getValue().getVolumeId());
             }
         }
         if (!backupsToRemove.isEmpty()) {
@@ -184,7 +176,7 @@ public class RetentionServiceImpl implements RetentionService {
     private Map<String, Set<BackupEntry>> getBackups() {
         Map<String, Set<BackupEntry>> backupsSortedByCreationTime = new HashMap<>();
 
-        for (BackupEntry backupEntry : backupRepository.findAll(instanceId)) {
+        for (BackupEntry backupEntry : backupRepository.findAll()) {
             Set<BackupEntry> backups = backupsSortedByCreationTime.get(backupEntry.getVolumeId());
             if (backups == null) {
                 backups = new TreeSet<>(backupComparatorByCreationTime);
@@ -199,7 +191,7 @@ public class RetentionServiceImpl implements RetentionService {
     private Map<String, RetentionEntry> getRetentions() {
         Map<String, RetentionEntry> retentionEntryMap = new HashMap<>();
 
-        for (RetentionEntry entry : retentionRepository.findByInstanceId(instanceId)) {
+        for (RetentionEntry entry : retentionRepository.findAll()) {
             retentionEntryMap.put(entry.getVolumeId(), entry);
         }
 
