@@ -1,43 +1,44 @@
-package com.sungardas.enhancedsnapshots.tasks;
-
-import com.amazonaws.AmazonClientException;
-import com.amazonaws.services.ec2.model.Instance;
-import com.amazonaws.services.ec2.model.Snapshot;
-import com.amazonaws.services.ec2.model.Volume;
-import com.sungardas.enhancedsnapshots.aws.dynamodb.model.BackupEntry;
-import com.sungardas.enhancedsnapshots.aws.dynamodb.model.BackupState;
-import com.sungardas.enhancedsnapshots.aws.dynamodb.model.TaskEntry;
-import com.sungardas.enhancedsnapshots.aws.dynamodb.model.WorkerConfiguration;
-import com.sungardas.enhancedsnapshots.aws.dynamodb.repository.BackupRepository;
-import com.sungardas.enhancedsnapshots.aws.dynamodb.repository.TaskRepository;
-import com.sungardas.enhancedsnapshots.dto.CopyingTaskProgressDto;
-import com.sungardas.enhancedsnapshots.exception.EnhancedSnapshotsInterruptedException;
-import com.sungardas.enhancedsnapshots.service.*;
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Profile;
-import org.springframework.context.annotation.Scope;
-import org.springframework.stereotype.Component;
+package com.sungardas.enhancedsnapshots.tasks.executors;
 
 import java.io.File;
 import java.io.IOException;
+import java.text.MessageFormat;
 import java.util.concurrent.TimeUnit;
+
+import com.amazonaws.services.ec2.model.Instance;
+import com.amazonaws.services.ec2.model.Snapshot;
+import com.amazonaws.services.ec2.model.Volume;
+import com.amazonaws.services.ec2.model.VolumeType;
+import com.sungardas.enhancedsnapshots.aws.dynamodb.model.BackupEntry;
+import com.sungardas.enhancedsnapshots.aws.dynamodb.model.BackupState;
+import com.sungardas.enhancedsnapshots.aws.dynamodb.model.TaskEntry;
+import com.sungardas.enhancedsnapshots.aws.dynamodb.repository.BackupRepository;
+import com.sungardas.enhancedsnapshots.aws.dynamodb.repository.TaskRepository;
+import com.sungardas.enhancedsnapshots.components.ConfigurationMediator;
+import com.sungardas.enhancedsnapshots.dto.CopyingTaskProgressDto;
+import com.sungardas.enhancedsnapshots.exception.EnhancedSnapshotsException;
+import com.sungardas.enhancedsnapshots.exception.EnhancedSnapshotsInterruptedException;
+import com.sungardas.enhancedsnapshots.service.AWSCommunicationService;
+import com.sungardas.enhancedsnapshots.service.NotificationService;
+import com.sungardas.enhancedsnapshots.service.RetentionService;
+import com.sungardas.enhancedsnapshots.service.SnapshotService;
+import com.sungardas.enhancedsnapshots.service.StorageService;
+import com.sungardas.enhancedsnapshots.service.TaskService;
+
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.Profile;
+import org.springframework.stereotype.Service;
 
 import static com.sungardas.enhancedsnapshots.aws.dynamodb.model.TaskEntry.TaskEntryStatus.ERROR;
 import static com.sungardas.enhancedsnapshots.aws.dynamodb.model.TaskEntry.TaskEntryStatus.RUNNING;
 import static java.lang.String.format;
 
-@Component
-@Scope("prototype")
+@Service("awsBackupVolumeTaskExecutor")
 @Profile("prod")
-public class AWSBackupVolumeTask implements BackupTask {
-    private static final Logger LOG = LogManager
-            .getLogger(AWSBackupVolumeTask.class);
-
-    @Value("${sungardas.worker.configuration}")
-    private String configurationId;
+public class AWSBackupVolumeTaskExecutor implements TaskExecutor {
+    private static final Logger LOG = LogManager.getLogger(AWSBackupVolumeTaskExecutor.class);
 
     @Autowired
     private TaskRepository taskRepository;
@@ -48,16 +49,14 @@ public class AWSBackupVolumeTask implements BackupTask {
     @Autowired
     private BackupRepository backupRepository;
 
-
     @Autowired
     private SnapshotService snapshotService;
 
     @Autowired
     private AWSCommunicationService awsCommunication;
-    private TaskEntry taskEntry;
 
     @Autowired
-    private ConfigurationService configurationService;
+    private ConfigurationMediator configurationMediator;
 
     @Autowired
     private RetentionService retentionService;
@@ -68,48 +67,31 @@ public class AWSBackupVolumeTask implements BackupTask {
     @Autowired
     private TaskService taskService;
 
-    private WorkerConfiguration configuration;
-
-    public void setTaskEntry(TaskEntry taskEntry) {
-        this.taskEntry = taskEntry;
-    }
-
-    public void execute() {
+    public void execute(TaskEntry taskEntry) {
         String volumeId = taskEntry.getVolume();
+        Volume tempVolume = null;
         try {
-            if (Thread.interrupted()) {
-                throw new EnhancedSnapshotsInterruptedException("Task interrupted");
-            }
+            checkThreadInterruption(taskEntry);
             notificationService.notifyAboutTaskProgress(taskEntry.getId(), "Starting backup task", 0);
-            configuration = configurationService.getWorkerConfiguration();
 
-            LOG.info(format(
-                    "AWSBackupVolumeTask: Starting backup process for volume %s",
-                    volumeId));
-            LOG.info("Task " + taskEntry.getId()
-                    + ": Change task state to 'inprogress'");
+            LOG.info("Starting backup process for volume {}", volumeId);
+            LOG.info("{} task state was changed to 'in progress'", taskEntry.getId());
             taskEntry.setStatus(RUNNING.getStatus());
             taskRepository.save(taskEntry);
 
-            Volume tempVolume = null;
-            String attachedDeviceName = null;
+            checkThreadInterruption(taskEntry);
 
-            if (Thread.interrupted()) {
-                throw new EnhancedSnapshotsInterruptedException("Task interrupted");
-            }
             notificationService.notifyAboutTaskProgress(taskEntry.getId(), "Preparing temp volume", 5);
-            tempVolume = createAndAttachBackupVolume(volumeId,
-                    configuration.getConfigurationId());
+
+            tempVolume = createAndAttachBackupVolume(volumeId, configurationMediator.getConfigurationId(), taskEntry);
             try {
                 TimeUnit.MINUTES.sleep(1);
             } catch (InterruptedException e1) {
                 e1.printStackTrace();
             }
-            if (Thread.interrupted()) {
-                throw new EnhancedSnapshotsInterruptedException("Task interrupted");
-            }
+            checkThreadInterruption(taskEntry);
             notificationService.notifyAboutTaskProgress(taskEntry.getId(), "Checking volume", 10);
-            attachedDeviceName = storageService.detectFsDevName(tempVolume);
+            String attachedDeviceName = storageService.detectFsDevName(tempVolume);
 
             String backupDate = String.valueOf(System.currentTimeMillis());
 
@@ -123,23 +105,20 @@ public class AWSBackupVolumeTask implements BackupTask {
 	    if (volumeType.equals("standard")) volumeType = "gp2";
             String backupFileName = volumeId + "." + backupDate + "." + volumeType + "." + iops + ".backup";
 
-            BackupEntry backup = new BackupEntry(volumeId, backupFileName, backupDate, "", BackupState.INPROGRESS,
-                    configuration.getConfigurationId(), snapshotId, volumeType, iops, sizeGib);
-            if (Thread.interrupted()) {
-                throw new EnhancedSnapshotsInterruptedException("Task interrupted");
-            }
+            BackupEntry backup = new BackupEntry(volumeId, backupFileName, backupDate, "", BackupState.INPROGRESS, snapshotId, volumeType, iops, sizeGib);
+            checkThreadInterruption(taskEntry);
             notificationService.notifyAboutTaskProgress(taskEntry.getId(), "Copying...", 15);
             boolean backupStatus = false;
             try {
                 String source = attachedDeviceName;
                 LOG.info("Starting copying: " + source + " to:" + backupFileName);
                 CopyingTaskProgressDto dto = new CopyingTaskProgressDto(taskEntry.getId(), 15, 80, Long.parseLong(backup.getSizeGiB()));
-                storageService.javaBinaryCopy(source, configuration.getSdfsMountPoint() + backupFileName, dto);
+                storageService.javaBinaryCopy(source, configurationMediator.getSdfsMountPoint() + backupFileName, dto);
                 backupStatus = true;
             } catch (IOException | InterruptedException e) {
                 LOG.fatal(format("Backup of volume %s failed", volumeId));
                 LOG.fatal(e);
-                File brocken = new File(configuration.getSdfsMountPoint() + backupFileName);
+                File brocken = new File(configurationMediator.getSdfsMountPoint() + backupFileName);
                 if (brocken.exists()) {
                     if (brocken.delete()) {
                         LOG.info("Broken backup {} was deleted", brocken.getName());
@@ -148,30 +127,22 @@ public class AWSBackupVolumeTask implements BackupTask {
                     }
                 }
             }
-            if (Thread.interrupted()) {
-                throw new EnhancedSnapshotsInterruptedException("Task interrupted");
-            }
+            checkThreadInterruption(taskEntry);
             notificationService.notifyAboutTaskProgress(taskEntry.getId(), "Detaching temp volume", 80);
             LOG.info("Detaching volume: {}", tempVolume.getVolumeId());
             awsCommunication.detachVolume(tempVolume);
-            if (Thread.interrupted()) {
-                throw new EnhancedSnapshotsInterruptedException("Task interrupted");
-            }
+            checkThreadInterruption(taskEntry);
             notificationService.notifyAboutTaskProgress(taskEntry.getId(), "Deleting temp volume", 85);
             LOG.info("Deleting temporary volume: {}", tempVolume.getVolumeId());
             awsCommunication.deleteVolume(tempVolume);
-            if (Thread.interrupted()) {
-                throw new EnhancedSnapshotsInterruptedException("Task interrupted");
-            }
+            checkThreadInterruption(taskEntry);
             if (backupStatus) {
-                long backupSize = storageService.getSize(configuration.getSdfsMountPoint() + backupFileName);
-                long backupCreationtime = storageService.getBackupCreationTime(configuration.getSdfsMountPoint() + backupFileName);
+                long backupSize = storageService.getSize(configurationMediator.getSdfsMountPoint() + backupFileName);
+                long backupCreationtime = storageService.getBackupCreationTime(configurationMediator.getSdfsMountPoint() + backupFileName);
                 LOG.info("Backup creation time: {}", backupCreationtime);
                 LOG.info("Backup size: {}", backupSize);
 
-                if (Thread.interrupted()) {
-                    throw new EnhancedSnapshotsInterruptedException("Task interrupted");
-                }
+                checkThreadInterruption(taskEntry);
                 LOG.info("Put backup entry to the Backup List: {}", backup.toString());
                 notificationService.notifyAboutTaskProgress(taskEntry.getId(), "Backup complete", 90);
                 backup.setState(BackupState.COMPLETED.getState());
@@ -181,26 +152,22 @@ public class AWSBackupVolumeTask implements BackupTask {
                 LOG.info(format("Backup process for volume %s finished successfully ", volumeId));
                 LOG.info("Task " + taskEntry.getId() + ": Delete completed task:" + taskEntry.getId());
                 LOG.info("Cleaning up previously created snapshots");
-                LOG.info("Storing snapshot data: [{},{},{}]", volumeId, snapshotId, configurationId);
+                LOG.info("Storing snapshot data: [{},{},{}]", volumeId, snapshotId, configurationMediator.getConfigurationId());
 
-                String previousSnapshot = snapshotService.getSnapshotId(volumeId, configurationId);
+                String previousSnapshot = snapshotService.getSnapshotId(volumeId);
                 if (previousSnapshot != null) {
-                    if (Thread.interrupted()) {
-                        throw new EnhancedSnapshotsInterruptedException("Task interrupted");
-                    }
+                    checkThreadInterruption(taskEntry);
                     notificationService.notifyAboutTaskProgress(taskEntry.getId(), "Deleting previous snapshot", 95);
                     LOG.info("Deleting previous snapshot {}", previousSnapshot);
                     awsCommunication.deleteSnapshot(previousSnapshot);
                 }
 
-                snapshotService.saveSnapshot(volumeId, configurationId, snapshotId);
+                snapshotService.saveSnapshot(volumeId, snapshotId);
 
 
                 taskService.complete(taskEntry);
                 LOG.info("Task completed.");
-                if (Thread.interrupted()) {
-                    throw new EnhancedSnapshotsInterruptedException("Task interrupted");
-                }
+                checkThreadInterruption(taskEntry);
                 notificationService.notifyAboutTaskProgress(taskEntry.getId(), "Task complete", 100);
                 retentionService.apply();
             } else {
@@ -208,51 +175,65 @@ public class AWSBackupVolumeTask implements BackupTask {
                 taskEntry.setStatus(ERROR.toString());
                 taskRepository.save(taskEntry);
             }
-        } catch (AmazonClientException e) {
-            LOG.error(format("Backup process for volume %s failed ", volumeId));
-            LOG.error(e);
+        } catch (Exception e) {
+            // TODO: add user notification about task failure
+            LOG.error("Backup process for volume {} failed ", volumeId, e);
             taskEntry.setStatus(ERROR.toString());
             taskRepository.save(taskEntry);
+
+            // clean up
+            if (tempVolume != null && awsCommunication.volumeExists(tempVolume.getVolumeId())) {
+                tempVolume = awsCommunication.syncVolume(tempVolume);
+                if (tempVolume.getAttachments().size() != 0) {
+                    awsCommunication.detachVolume(tempVolume);
+                }
+                awsCommunication.deleteVolume(tempVolume);
+            }
         }
     }
 
-    private Volume createAndAttachBackupVolume(String volumeId,
-                                               String instanceId) {
+    private Volume createAndAttachBackupVolume(String volumeId, String instanceId, TaskEntry taskEntry) {
         Instance instance = awsCommunication.getInstance(instanceId);
         if (instance == null) {
-            LOG.error("\nCan't get access to " + instanceId + " instance");
-
+            LOG.error("Can't get access to {} instance" + instanceId);
+            throw new AWSBackupVolumeException(MessageFormat.format("Can't get access to {} instance", instanceId));
         }
-        LOG.info("\ninst:" + instance);
 
         // create snapshot for AMI
         Volume volumeSrc = awsCommunication.getVolume(volumeId);
         if (volumeSrc == null) {
-            LOG.error("\nCan't get access to " + volumeId + " volume");
-
+            LOG.error("Can't get access to {} volume", volumeId);
+            throw new AWSBackupVolumeException(MessageFormat.format("Can't get access to {} volume", volumeId));
         }
 
-        Snapshot snapshot = awsCommunication
-                .waitForCompleteState(awsCommunication
-                        .createSnapshot(volumeSrc));
-        LOG.info("\nSnapshotEntry created. Check snapshot data:\n"
-                + snapshot.toString());
+        Snapshot snapshot = awsCommunication.waitForCompleteState(awsCommunication.createSnapshot(volumeSrc));
+        LOG.info("SnapshotEntry created: {}", snapshot.toString());
 
         // create volume
-        String instanceAvailabilityZone = instance.getPlacement()
-                .getAvailabilityZone();
-        Volume volumeDest = awsCommunication
-                .waitForAvailableState(awsCommunication
-                        .createVolumeFromSnapshot(snapshot,
-                                instanceAvailabilityZone));
-        LOG.info("\nVolume created. Check volume data:\n"
-                + volumeDest.toString());
+        String instanceAvailabilityZone = instance.getPlacement().getAvailabilityZone();
+        Volume volumeDest = awsCommunication.waitForAvailableState(awsCommunication.createVolumeFromSnapshot(snapshot.getSnapshotId(),
+                instanceAvailabilityZone, VolumeType.fromValue(taskEntry.getTempVolumeType()), taskEntry.getTempVolumeIopsPerGb()));
+        LOG.info("Volume created: {}", volumeDest.toString());
 
         // create temporary tag
         awsCommunication.createTemporaryTag(volumeDest.getVolumeId(),volumeSrc.getVolumeId());
 
         // mount AMI volume
         awsCommunication.attachVolume(instance, volumeDest);
+
         return awsCommunication.syncVolume(volumeDest);
+    }
+
+    private void checkThreadInterruption(TaskEntry taskEntry) {
+        if (Thread.interrupted()) {
+            LOG.info("Backup task {} was interrupted.", taskEntry.getId());
+            throw new EnhancedSnapshotsInterruptedException("Task interrupted");
+        }
+    }
+
+    public class AWSBackupVolumeException extends EnhancedSnapshotsException {
+        public AWSBackupVolumeException(String message) {
+            super(message);
+        }
     }
 }
