@@ -1,4 +1,6 @@
-package com.sungardas.enhancedsnapshots.tasks;
+package com.sungardas.enhancedsnapshots.tasks.executors;
+
+import java.util.concurrent.TimeUnit;
 
 import com.sungardas.enhancedsnapshots.aws.dynamodb.model.BackupEntry;
 import com.sungardas.enhancedsnapshots.aws.dynamodb.model.BackupState;
@@ -10,22 +12,19 @@ import com.sungardas.enhancedsnapshots.exception.EnhancedSnapshotsInterruptedExc
 import com.sungardas.enhancedsnapshots.service.NotificationService;
 import com.sungardas.enhancedsnapshots.service.RetentionService;
 import com.sungardas.enhancedsnapshots.service.TaskService;
+
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Profile;
-import org.springframework.context.annotation.Scope;
-import org.springframework.stereotype.Component;
-
-import java.util.concurrent.TimeUnit;
+import org.springframework.stereotype.Service;
 
 import static com.sungardas.enhancedsnapshots.aws.dynamodb.model.TaskEntry.TaskEntryStatus.RUNNING;
 
-@Component
-@Scope("prototype")
+@Service("awsBackupVolumeTaskExecutor")
 @Profile("dev")
-public class BackupFakeTask implements BackupTask {
-	private static final Logger LOG = LogManager.getLogger(BackupFakeTask.class);
+public class BackupFakeTaskExecutor implements TaskExecutor {
+	private static final Logger LOG = LogManager.getLogger(BackupFakeTaskExecutor.class);
     
     @Autowired
 	private TaskRepository taskRepository;
@@ -41,15 +40,8 @@ public class BackupFakeTask implements BackupTask {
     @Autowired
     private TaskService taskService;
 
-    private TaskEntry taskEntry;
-
-    
-    public void setTaskEntry(TaskEntry taskEntry) {
-    	this.taskEntry= taskEntry;
-    }
-
     @Override
-    public void execute() {
+    public void execute(TaskEntry taskEntry) {
         LOG.info("Task " + taskEntry.getId() + ": Change task state to 'inprogress'");
         notificationService.notifyAboutTaskProgress(taskEntry.getId(), "Starting delete task", 0);
 
@@ -62,8 +54,8 @@ public class BackupFakeTask implements BackupTask {
         String volumeId = taskEntry.getVolume();
         String filename = volumeId + "." + timestamp + ".backup";
         notificationService.notifyAboutTaskProgress(taskEntry.getId(), "Checking volume", 60);
-        BackupEntry backup = new BackupEntry(taskEntry.getVolume(), filename, timestamp, "123456789", BackupState.COMPLETED, taskEntry.getInstanceId(),
-        		"snap-00100110","gp2","3000", "10");
+        BackupEntry backup = new BackupEntry(taskEntry.getVolume(), filename, timestamp, "123456789", BackupState.COMPLETED,
+                "snap-00100110","gp2","3000", "10");
         LOG.info("Task " + taskEntry.getId() + ":put backup info'");
         backupRepository.save(backup);
 
