@@ -1,14 +1,17 @@
 package com.sungardas.init;
 
+import javax.annotation.PostConstruct;
+
 import com.amazonaws.AmazonClientException;
 import com.sungardas.enhancedsnapshots.aws.dynamodb.model.User;
 import com.sungardas.enhancedsnapshots.aws.dynamodb.repository.UserRepository;
 import com.sungardas.enhancedsnapshots.dto.InitConfigurationDto;
+import com.sungardas.enhancedsnapshots.dto.converter.BucketNameValidationDTO;
 import com.sungardas.enhancedsnapshots.exception.ConfigurationException;
 import com.sungardas.enhancedsnapshots.exception.EnhancedSnapshotsException;
 import com.sungardas.enhancedsnapshots.rest.RestAuthenticationFilter;
 import com.sungardas.enhancedsnapshots.rest.filters.FilterProxy;
-import com.sungardas.enhancedsnapshots.service.SharedDataService;
+
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.springframework.beans.BeansException;
@@ -20,11 +23,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.context.support.XmlWebApplicationContext;
 
-import javax.annotation.PostConstruct;
-import java.util.Arrays;
-
 import static org.springframework.http.HttpStatus.INTERNAL_SERVER_ERROR;
-import static org.springframework.http.HttpStatus.OK;
 
 
 @RestController
@@ -36,10 +35,8 @@ class InitController implements ApplicationContextAware {
     private FilterProxy filterProxy;
 
     @Autowired
-    private CredentialsService credentialsService;
-
-    @Autowired
-    private SharedDataService sharedDataService;
+    private InitConfigurationService initConfigurationService;
+    private InitConfigurationDto configurationDto;
 
     @Autowired
     private XmlWebApplicationContext applicationContext;
@@ -50,12 +47,12 @@ class InitController implements ApplicationContextAware {
     private void init() {
         // check that aws credentials are provided
         // try to authenticate as real admin user
-        if (credentialsService.isAwsPropertyFileExists()) {
-            LOG.info("Valid aws credentials were provided.");
+        if (initConfigurationService.propertyFileExists()) {
+            LOG.info("System is already configured.");
+            initConfigurationService.syncSettingsInDbAndConfigFile();
             refreshContext();
         } else {
-            credentialsService.configureAWSLogAgent();
-
+            initConfigurationService.configureAWSLogAgent();
         }
     }
 
@@ -67,6 +64,11 @@ class InitController implements ApplicationContextAware {
         return exception;
     }
 
+    //TODO: This should be removed, for dev mode only
+    @RequestMapping(value = "/configuration/awscreds", method = RequestMethod.GET)
+    public ResponseEntity<String> getAwsCredentialsInfo() {
+        return new ResponseEntity<>("{\"contains\": true}", HttpStatus.OK);
+    }
 
     @ExceptionHandler(value = {Exception.class, AmazonClientException.class})
     @ResponseBody
@@ -83,63 +85,53 @@ class InitController implements ApplicationContextAware {
         }
         // no aws credentials are provided
         // try to authenticate as default user admin@enhancedsnapshots:<instance-id>
-        else if (credentialsService.checkDefaultUser(user.getEmail(), user.getPassword())) {
+        else if (initConfigurationService.checkDefaultUser(user.getEmail(), user.getPassword())) {
             return new ResponseEntity<>("{ \"role\":\"configurator\" }", HttpStatus.OK);
         } else {
             return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
         }
     }
 
-    @RequestMapping(value = "/configuration/awscreds", method = RequestMethod.POST)
-    public ResponseEntity<String> setAwsCredential(@RequestBody CredentialsDto credentials) {
-        credentialsService.setCredentialsIfValid(credentials);
-        LOG.info("provided aws keys");
-        return new ResponseEntity<>(OK);
-    }
-
-    @RequestMapping(value = "/configuration/awscreds", method = RequestMethod.GET)
-    public ResponseEntity<String> getAwsCredentialsInfo() {
-        if (credentialsService.credentialsAreProvided()) {
-            return new ResponseEntity<>("{\"contains\": true}", HttpStatus.OK);
-        } else {
-            return new ResponseEntity<>("{\"contains\": false}", HttpStatus.OK);
-        }
-    }
-
-
     @RequestMapping(value = "/configuration/current", method = RequestMethod.GET)
     public ResponseEntity<InitConfigurationDto> getConfiguration() {
-	return new ResponseEntity<>(credentialsService.getInitConfigurationDto(), HttpStatus.OK);
+        return new ResponseEntity<>(getInitConfigurationDTO(), HttpStatus.OK);
     }
 
 
     @RequestMapping(value = "/configuration/current", method = RequestMethod.POST)
     public ResponseEntity<String> setConfiguration(@RequestBody ConfigDto config) {
-        if (credentialsService.areCredentialsValid()) {
-            InitConfigurationDto initConfigurationDto = credentialsService.getInitConfigurationDto();
-            if (!initConfigurationDto.getDb().isValid()) {
-                if (config.getUser() == null) {
-                    throw new ConfigurationException("Please create default user");
-                }
-                sharedDataService.setUser(config.getUser());
+        InitConfigurationDto initConfigurationDto = getInitConfigurationDTO();
+        if (!initConfigurationDto.getDb().isValid()) {
+            if (config.getUser() == null) {
+                throw new ConfigurationException("Please create default user");
             }
-            if (config.getUser() != null) {
-                sharedDataService.setUser(config.getUser());
-            }
-            initConfigurationDto.setS3(Arrays.asList(new InitConfigurationDto.S3(config.getBucketName(), false)));
-            sharedDataService.setInitConfigurationDto(initConfigurationDto);
-            credentialsService.storeCredentials();
-
-            try {
-                refreshContext();
-            } catch (Exception e) {
-                credentialsService.removeCredentials();
-                throw e;
-            }
-            return new ResponseEntity<>("", HttpStatus.OK);
-        } else {
-	    throw new ConfigurationException("AWS credentials invalid");
+            initConfigurationService.setUser(config.getUser());
         }
+        if (config.getUser() != null) {
+            initConfigurationService.setUser(config.getUser());
+        }
+        initConfigurationService.validateVolumeSize(config.getVolumeSize());
+        try {
+            // we need to ensure before context refresh that provided bucket name is valid
+            initConfigurationService.createBucket(config.getBucketName());
+        } catch (IllegalArgumentException e) {
+            LOG.warn("Failed to create bucket {}", config.getBucketName());
+            return new ResponseEntity<>(e.getMessage(), HttpStatus.UNPROCESSABLE_ENTITY);
+        }
+        initConfigurationService.storePropertiesEditableFromConfigFile();
+        initConfigurationService.createDBAndStoreSettings(config);
+        try {
+            refreshContext();
+        } catch (Exception e) {
+            initConfigurationService.removeProperties();
+            throw e;
+        }
+        return new ResponseEntity<>("", HttpStatus.OK);
+    }
+
+    @RequestMapping(value = "/configuration/bucket/{name:.+}", method = RequestMethod.GET)
+    public ResponseEntity<BucketNameValidationDTO> validateBucketName(@PathVariable("name") String bucketName) {
+        return new ResponseEntity<>(initConfigurationService.validateBucketName(bucketName), HttpStatus.OK);
     }
 
     @Override
@@ -156,16 +148,25 @@ class InitController implements ApplicationContextAware {
         // enabling auth filter
         RestAuthenticationFilter filter = applicationContext.getBean(RestAuthenticationFilter.class);
         filter.setUserRepository(applicationContext.getBean(UserRepository.class));
-        filter.setInstanceId(credentialsService.getInstanceId());
+        filter.setInstanceId(initConfigurationService.getInstanceId());
         filterProxy.setFilter(filter);
 
         LOG.info("Context refreshed successfully.");
         CONTEXT_REFRESH_IN_PROCESS = false;
     }
 
-    private static class ConfigDto {
+    static class ConfigDto {
         private User user;
         private String bucketName;
+        private int volumeSize;
+
+        public int getVolumeSize() {
+            return volumeSize;
+        }
+
+        public void setVolumeSize(final int volumeSize) {
+            this.volumeSize = volumeSize;
+        }
 
         public User getUser() {
             return user;
@@ -182,5 +183,12 @@ class InitController implements ApplicationContextAware {
         public void setBucketName(String bucketName) {
             this.bucketName = bucketName;
         }
+    }
+
+    private InitConfigurationDto getInitConfigurationDTO() {
+        if (configurationDto == null) {
+            configurationDto = initConfigurationService.getInitConfigurationDto();
+        }
+        return configurationDto;
     }
 }
